backend/requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
blinker==1.5
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
dj-database-url==1.0.0
Django==3.2
django-cors-headers==3.13.0
django-on-heroku==1.1.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
docutils==0.19
drf-yasg==1.21.4
feedgenerator==2.0.0
filelock==3.8.0
Flask==2.2.2
Flask-RESTful==0.3.9
ghp-import==2.1.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.3
openapi-codec==1.3.2
packaging==21.3
pelican==4.8.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.9.4
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
rich==12.6.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
Werkzeug==2.2.2
whitenoise==6.2.0
